feat: Move useLoadImage file into cornerstoneHelper
