+++ packages/insight-viewer/src/modules/cornerstoneHelper/index.ts
@@ -1,10 +1,2 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-const cornerstone = require('cornerstone-core')
-
-export function init(element: HTMLDivElement): void {
-  cornerstone.enable(element)
-}
-
-export function dispose(element: HTMLDivElement): void {
-  cornerstone.disable(element)
-}
+export { init, dispose } from './lifeCycle'
+export { useLoadImage } from './useLoadImage'
